Backend/TryOn/fashn-vton-1.5/requirements.txt
# FASHN VTON (Try-On web app)
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.3.0
huggingface_hub>=0.20.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
tqdm>=4.65.0
einops>=0.6.0
onnxruntime-gpu>=1.14.0
matplotlib>=3.5.0
fashn-human-parser>=0.1.1
flask
werkzeug
